added pulsebuilding.py to be the backend of mainwindow
